pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0",
    "wheel",
    "cibuildwheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "c3s-event-attribution-tools"
version = "0.0.1"  # CI will replace this during nightly build
description = "C3S 451 Event attribution tools library for Python"
readme = "README.md"
requires-python = ">=3.7"
license = "Apache-2.0"
license-files = [
    "LICENSE",
]
authors = [
  { name = "Paul Weerheim", email = "paul@maris.nl" },
  { name = "Tjerk Krijger", email = "tjerk@maris.nl" },
  { name = "Robin Kooyman", email = "robin@maris.nl" },
  { name = "Jasper van den Barg", email = "jasper@maris.nl"},
  { name = "Daniel Arteaga", email = "daniel@maris.nl"}
]

classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent"
]

dependencies = [
  "requests>=2.0.0",
  "pandas>=2.0.0",
  "pyarrow >= 19.0.1",
  "matplotlib >= 3.0.0",
  "xarray >= 2025.1.0",
  "netcdf4 >= 1.7.2",
  "cdsapi >= 0.7.6",
  "geopandas",
  "plotly",
  "contextily",
  "regionmask",
  "cmocean",
  "beacon-api >= 1.1.3",
  "cartopy",
  "networkx",
  "scitools-iris; sys_platform == 'linux' or sys_platform == 'darwin'",
  "ecmwf-api-client >= 1.6.1; sys_platform == 'linux' or sys_platform == 'darwin'"
]

[tool.setuptools]
packages = ["c3s_event_attribution_tools"]  # OR use find if you prefer
include-package-data = true

[tool.setuptools.package-data]
c3s_event_attribution_tools = ["RobotoCondensed-Regular.ttf", "LogoLine_horizon_C3S.png"]

[tool.cibuildwheel]
# Build all CPython versions (skip PyPy, Python 3.6)
build = "cp3*-*"
skip = "cp36-* pp*"